backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # For PostgreSQL
# pymysql==1.1.0  # Uncomment if using MySQL
